Remove the SDKv2 provider and protocol muxing
All resources are now served by the framework provider, so the SDKv2
provider and the mux are no longer needed:

- main.go serves the framework provider directly over protocol v6 via
  providerserver.Serve (no tf5to6server/tf6muxserver).
- The shared Airflow client moves to a new dependency-light internal/client
  package (no terraform-plugin-sdk import; the SDKv2 logging transport is
  replaced with http.DefaultTransport).
- internal/provider (the SDKv2 provider + its mux test) is deleted.
- base_endpoint is now a plain Optional attribute (presence enforced in
  Configure); the env-conditional schema hack that kept the muxed schemas
  identical is gone.
- The acceptance-test factory serves the framework provider only.

terraform-plugin-sdk/v2 remains only as an indirect dependency of the test
/docs tooling (terraform-plugin-testing, tfplugindocs); no provider code
references it. Regenerated docs (base_endpoint moves to Optional).
