+++ docs/index.md
@@ -129,12 +129,9 @@ curl -X POST https://airflow-server.net/auth/token \
 <!-- schema generated by tfplugindocs -->
 ## Schema
 
-### Required
-
-- `base_endpoint` (String)
-
 ### Optional
 
+- `base_endpoint` (String)
 - `base_path` (String) Base path for Airflow API endpoints
 - `disable_ssl_verification` (Boolean) Disable SSL verification
 - `oauth2_token` (String, Sensitive) The oauth to use for API authentication
