+++ internal/fwprovider/resource_dag.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/apache/airflow-client-go/airflow"
-	"github.com/drfaust92/terraform-provider-airflow/internal/provider"
+	"github.com/drfaust92/terraform-provider-airflow/internal/client"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -28,7 +28,7 @@ func newDagResource() resource.Resource {
 }
 
 type dagResource struct {
-	config provider.ProviderConfig
+	config client.ProviderConfig
 }
 
 type dagResourceModel struct {
@@ -109,11 +109,11 @@ func (r *dagResource) Configure(_ context.Context, req resource.ConfigureRequest
 		return
 	}
 
-	cfg, ok := req.ProviderData.(provider.ProviderConfig)
+	cfg, ok := req.ProviderData.(client.ProviderConfig)
 	if !ok {
 		resp.Diagnostics.AddError(
 			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected provider.ProviderConfig, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+			fmt.Sprintf("Expected client.ProviderConfig, got: %T. Please report this issue to the provider developers.", req.ProviderData),
 		)
 		return
 	}
