+++ internal/fwprovider/resource_dag_run.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/apache/airflow-client-go/airflow"
-	"github.com/drfaust92/terraform-provider-airflow/internal/provider"
+	"github.com/drfaust92/terraform-provider-airflow/internal/client"
 	"github.com/hashicorp/terraform-plugin-framework-timeouts/resource/timeouts"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
@@ -31,7 +31,7 @@ func newDagRunResource() resource.Resource {
 }
 
 type dagRunResource struct {
-	config provider.ProviderConfig
+	config client.ProviderConfig
 }
 
 type dagRunResourceModel struct {
@@ -100,11 +100,11 @@ func (r *dagRunResource) Configure(_ context.Context, req resource.ConfigureRequ
 		return
 	}
 
-	cfg, ok := req.ProviderData.(provider.ProviderConfig)
+	cfg, ok := req.ProviderData.(client.ProviderConfig)
 	if !ok {
 		resp.Diagnostics.AddError(
 			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected provider.ProviderConfig, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+			fmt.Sprintf("Expected client.ProviderConfig, got: %T. Please report this issue to the provider developers.", req.ProviderData),
 		)
 		return
 	}
